Bug 1308436 - Don't convert the curve duration from double to float in the EventInsertionHelper to avoid losing precision. r=karlt
MozReview-Commit-ID: 9vazrYi3J8C

Differential Revision: https://phabricator.services.mozilla.com/D4097

+++ dom/media/webaudio/AudioParam.h
@@ -219,7 +219,7 @@ private:
                             AudioTimelineEvent::Type aType,
                             double aTime, float aValue,
                             double aTimeConstant = 0.0,
-                            float aDuration = 0.0,
+                            double aDuration = 0.0,
                             const float* aCurve = nullptr,
                             uint32_t aCurveLength = 0)
   {
